Removed shared package (as a real package) and just literally share the Typescript file
