+++ packages/common/README.md
@@ -1,0 +1,5 @@
+We just want to be able to share some common string definitions
+between both our actual stack and the client library.
+
+There is very little "code" here - so we are not worried about
+the fact these functions end up duplicated into both packages.
